requirements.txt
pytest
pytest-mock
pytest-watch
pep8
pyfunctional
